# Optimize `CheckNumericOnly`